backend/requirements.txt
# Web Framework
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
asyncpg==0.29.0  # PostgreSQL async driver
aiosqlite==0.19.0  # SQLite async driver

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Async Tasks
celery[redis]==5.3.6
APScheduler==3.10.4
redis==5.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1
requests==2.32.3

# HTML Parsing
beautifulsoup4==4.12.3
lxml==5.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# OpenAI Integration
openai==1.7.2

# Torrent Parsing
bencodepy==0.9.5

# Media Info
pymediainfo==6.1.0
guessit==3.8.0  # Media title parsing

# File Operations
aiofiles==23.2.1

# Monitoring
python-json-logger==2.0.7
